requirements.txt
Django==4.1
django-cors-headers==3.9.0
djangorestframework==3.13.1
psycopg2==2.9.3